+++ src/config/parse.rs
@@ -4,12 +4,17 @@ use coveralls_api::CiService;
 use log::error;
 use regex::Regex;
 use serde::de::{self, Deserializer};
+use std::borrow::Cow;
 use std::env;
 use std::fmt;
 use std::fs::create_dir_all;
 use std::path::PathBuf;
 use std::str::FromStr;
 use std::time::Duration;
+
+pub(super) fn get_string(args: &ArgMatches, key: &str) -> Option<String> {
+    args.value_of_lossy(key).map(Cow::into_owned).or(None)
+}
 
 pub(super) fn get_list(args: &ArgMatches, key: &str) -> Vec<String> {
     args.values_of_lossy(key).unwrap_or_else(Vec::new)
@@ -19,7 +24,7 @@ pub(super) fn get_line_cov(args: &ArgMatches) -> bool {
     let cover_lines = args.is_present("line");
     let cover_branches = args.is_present("branch");
 
-    cover_lines || !(cover_lines || cover_branches)
+    cover_lines || !cover_branches
 }
 
 pub(super) fn get_branch_cov(args: &ArgMatches) -> bool {
@@ -56,6 +61,10 @@ pub(super) fn default_manifest() -> PathBuf {
     let mut manifest = env::current_dir().unwrap();
     manifest.push("Cargo.toml");
     manifest.canonicalize().unwrap_or(manifest)
+}
+
+pub(super) fn get_target(args: &ArgMatches) -> Option<String> {
+    args.value_of("target").map(String::from)
 }
 
 pub(super) fn get_target_dir(args: &ArgMatches) -> Option<PathBuf> {
@@ -99,15 +108,15 @@ pub(super) fn get_report_uri(args: &ArgMatches) -> Option<String> {
 }
 
 pub(super) fn get_outputs(args: &ArgMatches) -> Vec<OutputFile> {
-    values_t!(args.values_of("out"), OutputFile).unwrap_or(vec![])
+    values_t!(args.values_of("out"), OutputFile).unwrap_or_else(|_| vec![])
 }
 
 pub(super) fn get_output_directory(args: &ArgMatches) -> Option<PathBuf> {
-    args.value_of("output-dir").map(|path| PathBuf::from(path))
+    args.value_of("output-dir").map(PathBuf::from)
 }
 
 pub(super) fn get_run_types(args: &ArgMatches) -> Vec<RunType> {
-    values_t!(args.values_of("run-types"), RunType).unwrap_or(vec![RunType::Tests])
+    values_t!(args.values_of("run-types"), RunType).unwrap_or_else(|_| vec![RunType::Tests])
 }
 
 pub(super) fn get_excluded(args: &ArgMatches) -> Vec<Regex> {