+++ src/report/html.rs
@@ -58,7 +58,7 @@ fn get_json(coverage_data: &TraceMap, context: Context) -> Result<String, RunErr
     }
 
     safe_json::to_string_safe(&report)
-        .map_err(|e| RunError::Html(format!("Report isn't serializable: {}", e.to_string())))
+        .map_err(|e| RunError::Html(format!("Report isn't serializable: {}", e)))
 }
 
 pub fn export(coverage_data: &TraceMap, config: &Config) -> Result<(), RunError> {
@@ -79,7 +79,7 @@ pub fn export(coverage_data: &TraceMap, config: &Config) -> Result<(), RunError>
         None => String::from("null"),
     };
 
-    let html_write = match write!(
+    match write!(
         file,
         r##"<!doctype html>
 <html>
@@ -107,5 +107,5 @@ pub fn export(coverage_data: &TraceMap, config: &Config) -> Result<(), RunError>
         Err(e) => return Err(RunError::Html(e.to_string())),
     };
 
-    Ok(html_write)
+    Ok(())
 }