+++ src/source_analysis/mod.rs
@@ -209,8 +209,8 @@ fn is_target_folder(entry: &DirEntry, root: &Path) -> bool {
 fn is_hidden(entry: &DirEntry) -> bool {
     entry
         .path()
-        .into_iter()
-        .any(|x| x.to_string_lossy().starts_with("."))
+        .iter()
+        .any(|x| x.to_string_lossy().starts_with('.'))
 }
 
 /// Returns a list of files and line numbers to ignore (not indexes!)
@@ -289,7 +289,7 @@ pub fn debug_printout(result: &HashMap<PathBuf, LineAnalysis>, config: &Config) 
 fn analyse_lib_rs(file: &Path, result: &mut HashMap<PathBuf, LineAnalysis>) {
     if let Ok(f) = File::open(file) {
         let read_file = BufReader::new(f);
-        if let Some(Ok(first)) = read_file.lines().nth(0) {
+        if let Some(Ok(first)) = read_file.lines().next() {
             if !(first.starts_with("pub") || first.starts_with("fn")) {
                 let file = file.to_path_buf();
                 if result.contains_key(&file) {
@@ -377,10 +377,13 @@ fn analyse_package(
 /// These are often things like close braces, semi colons that may regiser as
 /// false positives.
 fn find_ignorable_lines(content: &str, analysis: &mut LineAnalysis) {
+    lazy_static! {
+        static ref IGNORABLE: Regex = Regex::new(r"^((\s*///)|([\[\]\{\}\(\)\s;\?,/]*$))").unwrap();
+    }
     let lines = content
         .lines()
         .enumerate()
-        .filter(|&(_, x)| !x.chars().any(|x| !"(){}[]?;\t ,".contains(x)))
+        .filter(|&(_, x)| IGNORABLE.is_match(&x))
         .map(|(i, _)| i + 1)
         .collect::<Vec<usize>>();
     analysis.add_to_ignore(&lines);