+++ tests/data/example_test/examples/say_hello.rs
@@ -1,5 +1,9 @@
 use example_test::hello_to;
 
+#[test]
+fn testit() {
+    main()
+}
 
 fn main() {
     let _hello_ferris = hello_to("Ferris");