+++ tests/failures.rs
@@ -1,6 +1,6 @@
 use crate::utils::get_test_path;
-use cargo_tarpaulin::launch_tarpaulin;
 use cargo_tarpaulin::{config::Config, errors::RunError};
+use cargo_tarpaulin::{launch_tarpaulin, run};
 use std::env;
 
 #[test]
@@ -16,6 +16,24 @@ fn error_if_compilation_fails() {
     assert!(result.is_err());
 
     if let Err(RunError::TestCompile(_)) = result {
+    } else {
+        panic!("Expected a TestCompile error");
+    }
+}
+
+#[test]
+fn error_if_test_fails() {
+    let mut config = Config::default();
+    let test_dir = get_test_path("failing_test");
+    env::set_current_dir(&test_dir).unwrap();
+    config.manifest = test_dir;
+    config.manifest.push("Cargo.toml");
+
+    let result = run(&[config]);
+
+    assert!(result.is_err());
+
+    if let Err(RunError::TestFailed) = result {
     } else {
         panic!("Expected a TestCompile error");
     }